# src/app/docs/swagger-suno-api.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "suno-api",
    "description": "Use API to call the music generation service of Suno.ai and easily integrate it into agents like GPTs.",
    "version": "1.1.0"
  },
...
  "paths": {
    "/api/generate": {
      "post": {
        "summary": "Generate audio based on Prompt.",
        "description": "It will automatically fill in the lyrics.\n\n2 audio files will be generated for each request, consuming a total of 10 credits.\n\n`wait_audio` can be set to API mode:\n\n\u2022 By default, it is set to `false`, which indicates the background mode. It will only return audio task information, and you will need to call the get API to retrieve detailed audio information.\n\n\u2022 If set to `true`, it simulates synchronous mode. The API will wait for a maximum of 100s until the audio is generated, and will directly return the audio link and other information. Recommend using in GPTs and other agents.",
        "tags": ["default"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["prompt", "make_instrumental", "wait_audio"],
                "properties": {
                  "prompt": {
                    "type": "string",
                    "description": "Prompt",
                    "example": "A popular heavy metal song about war, sung by a deep-voiced male singer, slowly and melodiously. The lyrics depict the sorrow of people after the war."
                  },
                  "make_instrumental": {
                    "type": "boolean",
                    "description": "Whether to generate instrumental music",
                    "example": "false"
                  },
                  "model": {
                    "type": "string",
                    "description": "Model name ,default is chirp-v3-5",
                    "example": "chirp-v3-5|chirp-v3-0"
                  },
                  "wait_audio": {
                    "type": "boolean",
                    "description": "Whether to wait for music generation, default is false, directly return audio task information; set to true, will wait for up to 100s until the audio is generated.",
                    "example": "false"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": ["0", "1"],
                    "properties": [
                      {
                        "$ref": "#/components/schemas/audio_info"
                      },
                      {
                        "$ref": "#/components/schemas/audio_info"
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/generate_stems": {
      "post": {
        "summary": "Make stem tracks (separate audio and music track).",
        "description": "Make stem tracks (separate audio and music track).",
        "tags": ["default"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["audio_id"],
                "properties": {
                  "audio_id": {
                    "type": "string",
                    "description": "The ID of the song to generate stems for.",
                    "example": "e76498dc-6ab4-4a10-a19f-8a095790e28d"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/schemas/audio_info"
          }
        }
      }
    },
    "/api/generate_lyrics": {
      "post": {
        "summary": "Generate lyrics based on Prompt.",
        "description": "Generate lyrics based on Prompt.",
        "tags": ["default"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["prompt"],
                "properties": {
                  "prompt": {
                    "type": "string",
                    "description": "Prompt",
                    "example": "A soothing lullaby"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "text": {
                      "type": "string",
                      "description": "Lyrics"
                    },
                    "title": {
                      "type": "string",
                      "description": "music title"
                    },
                    "status": {
                      "type": "string",
                      "description": "Status"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/get": {
      "get": {
        "summary": "Get audio information",
        "description": "",
        "tags": ["default"],
        "parameters": [
          {
            "in": "query",
            "name": "ids",
            "description": "Audio IDs, separated by commas. Leave blank to return a list of all music.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "description": "Page number",
            "required": false,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": ["0", "1"],
                    "properties": [
                      {
                        "$ref": "#/components/schemas/audio_info"
                      },
                      {
                        "$ref": "#/components/schemas/audio_info"
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/get_aligned_lyrics": {
      "get": {
        "summary": "Get lyric alignment.",
        "description": "Get lyric alignment.",
        "tags": ["default"],
        "parameters": [
          {
            "name": "song_id",
            "in": "query",
            "required": true,
            "description": "Song ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/schemas/audio_info"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "audio_info": {
        "type": "object",
        "required": [
          "id",
          "title",
          "image_url",
          "lyric",
          "audio_url",
          "video_url",
          "created_at",
          "model_name",
          "status",
          "gpt_description_prompt",
          "prompt",
          "type",
          "tags"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "audio id"
          },
          "title": {
            "type": "string",
            "description": "music title"
          },
          "image_url": {
            "type": "string",
            "description": "music cover image"
          },
          "lyric": {
            "type": "string",
            "description": "music lyric"
          },
          "audio_url": {
            "type": "string",
            "description": "music download url"
          },
          "video_url": {
            "type": "string",
            "description": "Music video download link, can be used to share"
          },
          "created_at": {
            "type": "string",
            "description": "Create time"
          },
          "model_name": {
            "type": "string",
            "description": "suno model name, chirp-v3"
          },
          "status": {
            "type": "string",
            "description": "The generated states include submitted, queue, streaming, complete."
          },
          "gpt_description_prompt": {
            "type": "string",
            "description": "Simple mode on user input prompt, Suno will generate formal prompts, lyrics, etc."
          },
          "prompt": {
            "type": "string",
            "description": "The final prompt for executing the generation task, customized by the user in custom mode, automatically generated by Suno in simple mode."
          },
          "type": {
            "type": "string",
            "description": "Type"
          },
          "tags": {
            "type": "string",
            "description": "Music genre. User-provided in custom mode, automatically generated by Suno in simple mode."
          }
        },
        "title": "audio_info",
        "description": "Audio Info"
      }
    }
  }
}